+++ quail/processes/__init__.py
@@ -1,7 +1,9 @@
 from .wps_climdex_days import ClimdexDays
-from .wps_climdexInput import ClimdexInput
+from .wps_climdexInput_csv import ClimdexInputCSV
+from .wps_climdexInput_raw import ClimdexInputRaw
 
 processes = [
     ClimdexDays(),
-    ClimdexInput(),
+    ClimdexInputCSV(),
+    ClimdexInputRaw(),
 ]
